REF: Use boolean types for brainmasks whenever possible
Use boolean types for brainmasks whenever possible: makes the
implementation match the attribute documentation:
https://github.com/nipreps/nifreeze/blob/eebdceb0ead93abab6a68685a684ece92cdfc526/src/nifreeze/data/base.py#L82

Note that `nibabel` does not support at this time boolean types when
building a `NIfTI` image instance or when serializing data. Thus, the
related statements remain unchanged or have been adapted to use
`np.uint8`.

+++ test/test_data_dmri.py
@@ -205,7 +205,7 @@ def test_equality_operator(tmp_path, setup_random_dwi_data):
     dwi, brainmask, b0 = _dwi_data_to_nifti(
         dwi_dataobj,
         affine,
-        brainmask_dataobj,
+        brainmask_dataobj.astype(np.uint8),
         b0_dataobj,
     )
